* removes unused file
* refactors func organization/hierarchy
