+++ src/networks/ALGO.ts
@@ -38,6 +38,8 @@ export const AlgoInterface: ConnectorInterface = {
   searchForTransactions,
 };
 
+export default AlgoInterface;
+
 /** Fetch account details (assets) */
 async function getAccount(address: string): Promise<any> {
   const { account } = await fetchAccount(address);
